Add Rust core interfaces
I've tried to keep the whole core in pure Rust using `gix` for the Git
processing. The plan is to use these interfaces on the Python side so
that spaCy can be used for parts-of-speech tagging. The whole CLI might
be moved back to native Rust at some point (I might try and natively
load the spaCy models if possible or find a different POS pipeline).

Commit-type: feature
Scope: core