Fix performance issues with `nlp.pipe` under macOS
Specifying an `n_process > 1` for `nlp.pipe` seems to substantially
increase the load time for the spaCy models under macOS. This seems to
have something to do with sharing pickled state to the threads.

Even without this flag the pipeline is still able to process
repositories with thousands of commits in seconds. Reintroducing the
`n_process` flag conditionally for other platforms can be done easily if
the need arises.

Commit-type: fix
Scope: cli
Closes: #2