Fix TUI rendering when downloading POS models
spaCy (+ pip under the hood) write to `stdout` and `stderr` in a way
that disrupts the TUI from rich. To resolve this the `--quiet` flag gets
propagated down to pip and the spaCy output gets redirected to
`devnull`. This output might contain useful error information (although
one hopes that spaCy outputs this against `stderr` and not `stdout`). An
alternative might be to figure out how to redirect the output stream
against the `rich.console` instead.

Commit-type: fix
Scope: cli